scripts/text-sync/requirements.txt
phonenumbers>=8.13.0
anthropic>=0.30.0
python-dotenv>=1.0.0
